requirements.txt
fastapi~=0.120.2
pydantic~=2.12.3
SQLAlchemy~=2.0.44
passlib~=1.7.4
alembic~=1.17.1
PyJWT~=2.10.1
python-dotenv~=1.2.1